requirements.txt
### Requirements:
Django==2.2.5
Psycopg2==2.8.3
Pillow==6.1.0
django-optimized-image==0.2.0
django-sendgrid-v5
python-decouple==3.1
django-ckeditor==5.7.1
django-imagekit==4.0.2
django_compressor==2.2
django-libsass==0.7
django-sass-processor==0.7.4
pytest
pytest-django==3.7.0
pytest-cov
pytest-datafiles==2.0
mixer
Faker==2.0.4
django-jet==1.0.8
django-distill==1.6
django-admin-sortable2==0.7.5
beautifulsoup4
djangorestframework
